requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
greenlet==3.0.3
gTTS==2.5.1
idna==3.7
numpy==1.26.4
pandas==2.2.2
PyAudio==0.2.14
pycparser==2.22
pyFLAC==3.0.0
pymssql==2.3.0
pyodbc==5.1.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
requests==2.31.0
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.3
SQLAlchemy==2.0.30
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
